review comment: move matching_dim to views.py
