+++ torch_spyre/_inductor/pass_utils.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from typing import NamedTuple, Optional
+from typing import NamedTuple
 
 from sympy import Expr, Symbol
 
@@ -87,20 +87,3 @@ def device_coordinates(layout: FixedTiledLayout, dep: MemoryDep) -> list[sympy.E
         dep.ranges,
         dep.index,
     )
-
-
-def matching_dim(coords: list[sympy.Expr], expr: sympy.Expr) -> Optional[int]:
-    """
-    Given a coordinate array and an expression, determine if there is a unique
-    dimension in coords whose coordinate expression is exactly the one free variable
-    in the expression.  Return None if expr does not have exactly one free variable
-    or if there is not exactly one matching dimension in coords.
-    """
-    if len(expr.free_symbols) != 1:
-        return None
-    v = next(iter(expr.free_symbols))
-    dims = [d for d, e in enumerate(coords) if e == v]
-    if len(dims) != 1:
-        return None
-    else:
-        return dims[0]
